moving types to messages
