+++ packages/cli/src/nodehost.ts
@@ -31,13 +31,11 @@ import {
 import { tryReadText } from "../../core/src/fs"
 import {
     ServerManager,
-    LanguageModelConfiguration,
     LogLevel,
     UTF8Decoder,
     UTF8Encoder,
     RuntimeHost,
     setRuntimeHost,
-    ResponseStatus,
     AzureTokenResolver,
     ModelConfigurations,
     ModelConfiguration,
@@ -52,7 +50,7 @@ import { shellConfirm, shellInput, shellSelect } from "./input"
 import { shellQuote } from "../../core/src/shell"
 import { uniq } from "es-toolkit"
 import { PLimitPromiseQueue } from "../../core/src/concurrency"
-import { Project } from "../../core/src/server/messages"
+import { LanguageModelConfiguration, Project, ResponseStatus } from "../../core/src/server/messages"
 import { createAzureTokenResolver } from "./azuretoken"
 import {
     createAzureContentSafetyClient,
