+++ packages/cli/src/server.ts
@@ -16,8 +16,6 @@ import {
     serializeError,
 } from "../../core/src/error"
 import {
-    LanguageModelConfiguration,
-    ResponseStatus,
     ServerResponse,
     host,
     runtimeHost,
@@ -34,6 +32,8 @@ import {
     ChatCancel,
     LanguageModelConfigurationResponse,
     PromptScriptListResponse,
+    ResponseStatus,
+    LanguageModelConfiguration,
 } from "../../core/src/server/messages"
 import { envInfo } from "./info"
 import { LanguageModel } from "../../core/src/chat"
