+++ packages/core/src/server/client.ts
@@ -2,10 +2,8 @@ import { ChatCompletionsProgressReport } from "../chattypes"
 import { CLIENT_RECONNECT_DELAY, OPEN, RECONNECT } from "../constants"
 import { randomHex } from "../crypto"
 import { errorMessage } from "../error"
-import { GenerationResult } from "../server/messages"
+import { GenerationResult, LanguageModelConfiguration, ResponseStatus } from "../server/messages"
 import {
-    LanguageModelConfiguration,
-    ResponseStatus,
     ServerResponse,
     host,
 } from "../host"
