+++ packages/core/src/testhost.ts
@@ -7,7 +7,6 @@ import { readFile, writeFile } from "fs/promises"
 import { ensureDir } from "fs-extra"
 import {
     LogLevel,
-    LanguageModelConfiguration,
     ServerManager,
     UTF8Decoder,
     UTF8Encoder,
@@ -16,7 +15,6 @@ import {
     AzureTokenResolver,
     ModelConfigurations,
     ModelConfiguration,
-    ResponseStatus,
 } from "./host"
 import { TraceOptions } from "./trace"
 import {
@@ -31,7 +29,7 @@ import {
 } from "node:path"
 import { LanguageModel } from "./chat"
 import { NotSupportedError } from "./error"
-import { Project } from "./server/messages"
+import { LanguageModelConfiguration, Project, ResponseStatus } from "./server/messages"
 import { defaultModelConfigurations } from "./llms"
 
 // Function to create a frozen object representing Node.js path methods
