+++ packages/core/src/vectorsearch.ts
@@ -16,7 +16,7 @@ import type { EmbeddingsModel, EmbeddingsResponse } from "vectra/lib/types"
 import { createFetch, traceFetchPost } from "./fetch"
 import { JSONLineCache } from "./cache"
 import { EmbeddingCreateParams, EmbeddingCreateResponse } from "./chattypes"
-import { LanguageModelConfiguration } from "./host"
+import { LanguageModelConfiguration } from "./server/messages"
 import { getConfigHeaders } from "./openai"
 import { logVerbose, trimTrailingSlash } from "./util"
 import { TraceOptions } from "./trace"
