+++ packages/web/src/App.tsx
@@ -1,4 +1,4 @@
-// src/components/FormField.tsx
+/// <reference path="../../core/src/types/prompt_template.d.ts" />
 import React, {
     createContext,
     Dispatch,
